Fixing some bugs, and working on the provider-request , notifications, and payment scheme.

+++ src/provider-profile/prisma-provider-profile.repository.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { ForbiddenException, Injectable } from "@nestjs/common";
 import { PrismaService } from "src/infrastructure/prisma/prisma.service";
 import { CreateProviderProfileDto } from "./dto/create-provider-profile.dto";
 import { ProviderProfile } from "./entities/provider-profile.entity";
@@ -28,13 +28,31 @@ export class PrismaProviderProfileRepository implements ProviderProfileRepositor
       where: { isActive: true },
     }) as Promise<ProviderProfile[]>;
   }
+  findByUserId(userId: string): Promise<ProviderProfile | null> {
+    return this.prisma.providerProfile.findUnique({
+      where: { userId },
+    });
+  }
   findById(id: string): Promise<ProviderProfile | null> {
     return this.prisma.providerProfile.findUnique({
       where: { id },
-      include: { user: true, services: true, availability: true },
+      include: {
+        user: true,
+        services: true,
+        availability: true,
+      },
     }) as Promise<ProviderProfile | null>;
   }
   async deleteProvider(id: string): Promise<{ message: string }> {
+    const provider = await this.prisma.providerProfile.findUnique({
+      where: { id },
+      include: { bookings: true },
+    });
+    if (provider?.bookings && provider.bookings.length > 0) {
+      throw new ForbiddenException(
+        "There booking connection with this provider",
+      );
+    }
     await this.prisma.providerProfile.delete({
       where: { id },
     });
